+++ ComeTogetherApp/ComeTogetherApp/Beans/Transaction.cs
@@ -16,9 +16,9 @@ namespace ComeTogetherApp
         public string type { get; set; }
         public string receiverName { get; set; }
         public string senderName { get; set; }
-        public int amount { get; set; }
+        public double amount { get; set; }
 
-        public Transaction(string receiver, string sender, string type, int amount)
+        public Transaction(string receiver, string sender, string type, double amount)
         {
             this.receiver = receiver;
             this.sender = sender;
